refactor : SettingWinningNumber 네이밍 변경 및 생성자에서 winningNumSet 로직 수행
SettingWinningNumber -> SetWinNum(당첨번호설정)
winningNumSet 로직 생성자와 통합
테스트 코드 및 메인 출력코드 SetWinNum 수정
