+++ STEP_02/src/me/gimun/SetWinNum.java
@@ -3,15 +3,12 @@ package me.gimun;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SettingWinningNumber {
+public class SetWinNum {
     private List<Integer> winningLotto;
     private int winningBonusNum;
 
-    public SettingWinningNumber() {
-        this.winningLotto = new ArrayList<Integer>();
-    }
-
-    public void winningNumSet(List<Integer> lotto){
+    public SetWinNum(List<Integer> lotto) {
+        this.winningLotto = new ArrayList<>();
         //로또 길이가 7이 아니면 에러
         if(lotto.size() != 7){
             throw new IllegalArgumentException();
@@ -20,8 +17,8 @@ public class SettingWinningNumber {
         this.winningBonusNum = lotto.get(lotto.size()-1);
         lotto.remove(lotto.size()-1);
         this.winningLotto = lotto;
-
     }
+
 
     public List<Integer> getWinningLotto() {
         return winningLotto;
