+++ STEP_02/test/me/gimun/ConfirmWinningTest.java
@@ -4,8 +4,6 @@ import org.junit.jupiter.api.Test;
 
 import java.util.*;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 class ConfirmWinningTest {
 
     @Test
@@ -16,7 +14,6 @@ class ConfirmWinningTest {
         List<Integer> lotto = lottoGenerate.getLotto();
 
         //당첨번호 지정
-        SettingWinningNumber settingWinningNumber = new SettingWinningNumber();
         Set<Integer> set = new HashSet<>();
         set.add(3);
         set.add(10);
@@ -31,14 +28,14 @@ class ConfirmWinningTest {
 
         //보너스번호
         winningLotto.add(45);
-        settingWinningNumber.winningNumSet(winningLotto);
+        SetWinNum setWinNum = new SetWinNum(winningLotto);
 
         //당첨확인
         ConfirmWinning confirmWinning = new ConfirmWinning();
-        String result = confirmWinning.confirmLotto(settingWinningNumber.getWinningLotto(),settingWinningNumber.getWinningBonusNum(),lotto);
+        String result = confirmWinning.confirmLotto(setWinNum.getWinningLotto(), setWinNum.getWinningBonusNum(),lotto);
 
         System.out.println("내가 구매한 로또 번호 : " + lotto);
-        System.out.println("당첨번호 : " + settingWinningNumber.getWinningLotto() + " 보너스번호 : " + settingWinningNumber.getWinningBonusNum());
+        System.out.println("당첨번호 : " + setWinNum.getWinningLotto() + " 보너스번호 : " + setWinNum.getWinningBonusNum());
         System.out.println("당첨확인 : " + result);
     }
 }
